submit java-oop-ru/errors

+++ java-oop-ru/errors/src/main/java/exercise/App.java
@@ -4,7 +4,7 @@ package exercise;
 public class App {
     public static int printSquare(Circle circle) {
         try {
-            (int)(Circle.getSquare(circle));
+            Circle.getSquare(circle);
         } catch (NegativeRadiusException e) {
             System.out.println("Не удалось посчитать площадь");
         } finally {
